--- apps/wildfires/src/app/pages/home/homePage.tsx
@@ -1,5 +1,0 @@
-import { FiresSurvey } from './firesSurvey/FiresSurvey';
-
-export function HomePage() {
-  return <FiresSurvey />;
-}
