+++ apps/wildfires/src/app/pages/introduction/index.tsx
@@ -1,0 +1,5 @@
+import { FiresSurvey } from './firesSurvey/FiresSurvey';
+
+export function Introduction() {
+  return <FiresSurvey />;
+}
