+++ apps/wildfires/src/app/shared/components/hero/Hero.tsx
@@ -1,0 +1,35 @@
+import { ReactNode } from 'react';
+
+interface IHeroProps {
+  className?: string;
+  backgroundImage?: string;
+  children: ReactNode;
+}
+
+export default function Hero(props: IHeroProps) {
+  const { className = '', backgroundImage, children } = props;
+
+  const parentCss = [
+    'min-h-[75vh]',
+    'flex',
+    'items-center',
+    'justify-left',
+    'bg-cover',
+    'bg-no-repeat',
+    'bg-center',
+    'w-full',
+    className,
+  ].join(' ');
+  return (
+    <div
+      className={parentCss}
+      style={{
+        backgroundImage: backgroundImage
+          ? `url(${backgroundImage})`
+          : undefined,
+      }}
+    >
+      {children}
+    </div>
+  );
+}
